pyproject.toml
[project]
name = "microsoft-agentic-ai"
version = "0.1.0"
description = "Microsoft Agent Framework A2A Orchestrator with SQL Foundry Agent and Python Tool Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Microsoft Agent Framework (preview) - for client wrapper
    # Pre-release versions will be allowed via [tool.uv] configuration below
    "agent-framework",

    # A2A SDK - for both client and server
    "a2a-sdk>=0.3.0",

    # Azure AI SDKs - for agent implementation
    "azure-ai-projects>=1.0.0b11",
    "azure-ai-agents>=1.0.0",
    "azure-identity>=1.23.0",

    # Semantic Kernel - for Python Tool Agent
    "semantic-kernel>=1.22.0",

    # Web server dependencies - for A2A server
    "uvicorn>=0.34.2",
    "starlette>=0.35.0",
    "sse-starlette>=2.3.5",

    # HTTP client - for both client and server
    "httpx>=0.25.0",

    # Data science libraries - for Python Tool Agent visualizations
    "matplotlib>=3.8.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "seaborn>=0.13.0",

    # CLI and utilities
    "click>=8.0.0",
    "python-dotenv>=1.0.0",

    # Web UI
    "streamlit>=1.28.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include both agent packages
packages = ["sql_foundry_agent", "python_tool_agent"]

[tool.uv]
# Allow pre-release versions globally (equivalent to pip install --pre)
prerelease = "allow"
